validations & new car modifications
Added validations related to id field and added a new car object with more fields.

+++ src/app/admin/admin.component.ts
@@ -8,7 +8,7 @@ import { CarService } from '../car.service';
   styleUrls: ['./admin.component.css']
 })
 export class AdminComponent implements OnInit {
-  selectedCar:Car = { id:'15' , name:'Subaru' ,type:'SUV' };
+  selectedCar:Car = { id:15 , name:'Subaru' ,type:'SUV' };
   cars:Car[];
   newCar:Car;
   constructor(private carService:CarService) { 
@@ -19,29 +19,41 @@ export class AdminComponent implements OnInit {
     this.carService.getAllCars().subscribe((data: Car[]) => this.cars = data );
   }
   //save new car details
-  saveCartoDB(id:string,name:string,type:string):void{
+  saveCartoDB(id:number,name:string,type:string):void{
     console.log("in car save");
+    if(isNaN(id)) {
+      alert("Please enter a number in ID field");
+    }else{
     this.newCar = {
       id: id,
       name:name,
       type:type
     };
     this.carService.addCar(this.newCar).subscribe();
+    }
   }
   //Update car details
   updateCartoDB(car:Car):void{
+    if(isNaN(car.id)) {
+      alert("Please enter a number in ID field");
+    }else{
     this.carService.updateCar(this.selectedCar).subscribe();
+    }
   }
   //delete car details
-  deleteCarinDB(id:string):void{
-   this.carService.deleteCar(id).subscribe();
-    console.log("car deleted");
+  deleteCarinDB(id:number):void{
+    if(isNaN(id)) {
+      alert("Please enter a number in ID field");
+    }else{
+      this.carService.deleteCar(id).subscribe();
+    }
+      console.log("car deleted");
   }
   onSelect(car:Car):void{
     this.selectedCar = car;
   }
   clearAll(){
-    this.selectedCar.id="";
+    this.selectedCar.id=null;
     this.selectedCar.name="";
     this.selectedCar.type="";
   }
